Refactor: Consolidate feed and stream related classes in com.fauna.event.
