+++ src/main/java/com/fauna/client/FaunaClient.java
@@ -4,19 +4,21 @@ import com.fauna.codec.Codec;
 import com.fauna.codec.CodecProvider;
 import com.fauna.codec.DefaultCodecProvider;
 import com.fauna.codec.DefaultCodecRegistry;
+import com.fauna.event.FaunaStream;
+import com.fauna.event.FeedIterator;
 import com.fauna.exception.ClientException;
 import com.fauna.exception.ErrorHandler;
 import com.fauna.exception.FaunaException;
 import com.fauna.exception.ProtocolException;
 import com.fauna.exception.ServiceException;
-import com.fauna.feed.EventSource;
-import com.fauna.feed.FeedOptions;
-import com.fauna.feed.FeedPage;
+import com.fauna.event.EventSource;
+import com.fauna.event.FeedOptions;
+import com.fauna.event.FeedPage;
 import com.fauna.query.AfterToken;
 import com.fauna.query.QueryOptions;
 import com.fauna.response.QueryFailure;
-import com.fauna.stream.StreamRequest;
-import com.fauna.query.EventSourceResponse;
+import com.fauna.event.StreamRequest;
+import com.fauna.event.EventSourceResponse;
 import com.fauna.query.builder.Query;
 import com.fauna.response.QueryResponse;
 import com.fauna.response.QuerySuccess;
