+++ src/test/java/com/fauna/e2e/E2EFeedsTest.java
@@ -3,14 +3,14 @@ package com.fauna.e2e;
 import com.fauna.client.Fauna;
 import com.fauna.client.FaunaClient;
 import com.fauna.client.FaunaConfig;
-import com.fauna.client.FeedIterator;
+import com.fauna.event.FeedIterator;
 import com.fauna.e2e.beans.Product;
-import com.fauna.feed.EventSource;
-import com.fauna.feed.FeedOptions;
-import com.fauna.feed.FeedPage;
-import com.fauna.query.EventSourceResponse;
+import com.fauna.event.EventSource;
+import com.fauna.event.FeedOptions;
+import com.fauna.event.FeedPage;
+import com.fauna.event.EventSourceResponse;
 import com.fauna.response.QuerySuccess;
-import com.fauna.response.StreamEvent;
+import com.fauna.event.StreamEvent;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
@@ -25,7 +25,7 @@ import java.util.stream.Collectors;
 
 import static com.fauna.client.FaunaConfig.FaunaEndpoint.LOCAL;
 import static com.fauna.query.builder.Query.fql;
-import static com.fauna.response.StreamEvent.EventType.ERROR;
+import static com.fauna.event.StreamEvent.EventType.ERROR;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
