+++ src/test/java/com/fauna/e2e/E2EStreamingTest.java
@@ -3,14 +3,14 @@ package com.fauna.e2e;
 import com.fauna.client.Fauna;
 import com.fauna.client.FaunaClient;
 import com.fauna.client.FaunaConfig;
-import com.fauna.client.FaunaStream;
+import com.fauna.event.FaunaStream;
 import com.fauna.e2e.beans.Product;
 import com.fauna.exception.ClientException;
-import com.fauna.query.EventSourceResponse;
+import com.fauna.event.EventSourceResponse;
 import com.fauna.query.builder.Query;
 import com.fauna.response.QuerySuccess;
-import com.fauna.response.StreamEvent;
-import com.fauna.stream.StreamRequest;
+import com.fauna.event.StreamEvent;
+import com.fauna.event.StreamRequest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
